Extract Item.Import
